requirements.txt
psycopg2-binary
sqlalchemy
pydantic-sqlalchemy
flask
Flask-SQLAlchemy